Fixes a bug that was pointed out by awalsh
This bug would output mix the output of COMPREPLY
along with the files present in the CWD.
This was fixed by moving the function one level
before COMPREPLY gives garbage output.

I also changed the source code and removed
unnessary lines of code and duplicate lines of
code.

I added comments so that it is easier to understand
what the function actually does

Signed-off-by: Achilles Gaikwad <agaikwad@redhat.com>